refactor(viewer): split CorpusTab into components/admin/corpus
1,323-line tab becomes a folder: thin CorpusTab, CorpusList, NewCorpusForm,
CorpusFiles (flat table on DataTable) with CorpusFilesHeader and
UploadFailuresDialog, CopyFromScopeModal (flat table on DataTable), shared
key helpers + view toggle, and a useCorpusFiles hook owning the listing,
pending folders, selection, uploads and every admin mutation. Markup,
class names and strings unchanged.
